moved out header from the main component
