+++ Localize/lang/strings.json
@@ -1263,7 +1263,6 @@
   "Yua/4o": "Previous",
   "Yuu5CD": "Zoom out",
   "Yuxprm": "Welcome to the workflow assistant!",
-  "YwhRHw": "Back to connectors",
   "Yz9o1k": "Not connected.",
   "Z8BOCl": "No identities available",
   "Z8tBFS": "The workflow assistant is designed only to provide help and doesn't support workflow creation or editing.",
@@ -2557,7 +2556,6 @@
   "_Yua/4o.comment": "Button text for moving to the previous tab in the create workflow panel",
   "_Yuu5CD.comment": "Label to zoom the canvas out",
   "_Yuxprm.comment": "Chatbot greeting message from existing flow",
-  "_YwhRHw.comment": "Button text to return to connector browse view",
   "_Yz9o1k.comment": "Text to show that no connection is connected to the node",
   "_Z8BOCl.comment": "Placeholder warning for no identities available",
   "_Z8tBFS.comment": "Chatbot disclaimer message that workflow assistant can only provide help and not modify workflows",
