+++ libs/designer/src/lib/ui/mcp/panel/connectorPanel/ConnectorPanel.tsx
@@ -2,11 +2,17 @@ import { TemplateContent, TemplatesPanelFooter, type McpPanelTabProps } from '@m
 import { useMcpConnectorPanelTabs } from './usePanelTabs';
 import { useDispatch, useSelector } from 'react-redux';
 import type { AppDispatch, RootState } from './../../../../core/state/mcp/store';
-import { selectPanelTab } from './../../../../core/state/mcp/panel/mcpPanelSlice';
-import { DrawerBody, DrawerFooter, DrawerHeader } from '@fluentui/react-components';
+import { closePanel, selectPanelTab } from './../../../../core/state/mcp/panel/mcpPanelSlice';
+import { Button, DrawerBody, DrawerFooter, DrawerHeader, Text } from '@fluentui/react-components';
 import { useMcpPanelStyles } from '../styles';
+import { useIntl } from 'react-intl';
+import { bundleIcon, Dismiss24Filled, Dismiss24Regular } from '@fluentui/react-icons';
+import { useCallback } from 'react';
+
+const CloseIcon = bundleIcon(Dismiss24Filled, Dismiss24Regular);
 
 export const ConnectorPanelInner = () => {
+  const intl = useIntl();
   const dispatch = useDispatch<AppDispatch>();
   const panelTabs: McpPanelTabProps[] = useMcpConnectorPanelTabs();
   const { selectedTabId } = useSelector((state: RootState) => ({
@@ -21,15 +27,37 @@ export const ConnectorPanelInner = () => {
 
   const styles = useMcpPanelStyles();
 
+  const INTL_TEXT = {
+    title: intl.formatMessage({
+      id: 'SH50TJ',
+      defaultMessage: 'Add Connectors',
+      description: 'Title for connector selection panel',
+    }),
+    closeAriaLabel: intl.formatMessage({
+      id: 'kdCuJZ',
+      defaultMessage: 'Close panel',
+      description: 'Aria label for close button',
+    }),
+  };
+
+  const handleDismiss = useCallback(() => {
+    dispatch(closePanel());
+  }, [dispatch]);
   return (
     <div>
-      <DrawerHeader className={styles.header}>{'header placeholder'}</DrawerHeader>
+      <DrawerHeader className={styles.header}>
+        <div className={styles.headerContent}>
+          <Text size={600} weight="semibold" style={{ flex: 1 }}>
+            {INTL_TEXT.title}
+          </Text>
+          <Button appearance="subtle" icon={<CloseIcon />} onClick={handleDismiss} aria-label={INTL_TEXT.closeAriaLabel} />
+        </div>
+      </DrawerHeader>
       <DrawerBody>
         <TemplateContent tabs={panelTabs} selectedTab={selectedTabId} selectTab={onTabSelected} />
       </DrawerBody>
       {selectedTabProps?.footerContent && (
         <DrawerFooter className={styles.footer}>
-          DKFJSLKDFJSDLKFj
           <TemplatesPanelFooter {...selectedTabProps.footerContent} />
         </DrawerFooter>
       )}
